Coded the guts of turret.java
VERY preliminary code; need to work out the empirical data, add debug
print statements, review the math, etc.
Also removed the analogEncoder because it is connected directly to the
Talon and therefore not visible to RobotBuilder.  It is accessed through
the motor controller.
Added an ActivateTurret command (visible on SmartDashboard) to run the
system for 5 seconds as a test vehicle.  Eventually this will be the
default command for the subsystem when enabled.