Add initial Shooter, clean up Turret, add debug print support
Added the Shooter subsystem and some joystick controls for it.  Can be
called from an autonomous command as well, but the code to determine
shot speed based on vision target distance isn't implemented yet....
ShooterRPM is just a value in the SmartDashboard for the moment.  Added
some joystick controls to Turret and cleaned up some PID initialization
and unit conversions.  Added some global flags for debug print messages
(mostly for shooter right now).